Added a new function, ipynast_seqs, an iterator which yields (sequence, exit status) tuples after each candidate sequences is passed to pynast_seq. The majority of functionality in this function was copied from pynast_seqs, as pynast_seqs has now been refactored as a wrapper for ipynast_seqs which collects all output in lists to return.
This change was made as it may be important at times for pynast_seqs to be able to function as an iterator. For example, if the user passes in millions of sequences, the previous version would store the aligned sequences in memory until all sequences have been processed. Now, these will be written out to file as the file buffer fills up, potentially freeing large amounts of memory.

A side effect of this change is that scripts/pynast.main, which now calls ipynast_seqs, no longer needs to pass a status callback function as it has the opportunity to do status calculations directly after each sequence is aligned. Those changes were therefore added as well. Note that because all callers of ipynast_seqs will have this opportunity, ipynast_seqs does not take a status callback function. pynast_seqs, however, does still allow users to pass a status callback function.