Major changes to PyNAST, primarily involving the switch to uclust for sequence searching and as the default pairwise aligner. This change is expected to decrease the runtime of pynast, and yield better results (based on changes noted in Schloss 2009). Some brief notes about this are posted on the PyNAST blog (as Feature 1) at: http://wp.me/pMXzd-7
This revision marks the break from the goal of matching the results of command line NAST. PyNAST no longer strives to match those results - the goal of the project is now to generate better alignments. Users who are interested in matching the result of NAST should use PyNAST 1.0. 

At this time, sequences are passed one at a time to uclust for searching and aligning. This is very inefficent, and changes to follow very shortly will improve upon this. (The changes were piling up though, and needed to be checked in.)